Refactor transaction hash resolvers and update import structure
- Introduced a new `resolver.ts` file to define the `TxHashResolver` type for improved type safety.
- Moved individual transaction hash resolver functions for various chains (Cardano, Cosmos, EVM, Polkadot, Ripple, Solana, Sui, Ton, Tron, UTXO) into a dedicated `resolvers` directory, enhancing organization and maintainability.
- Updated import paths across the codebase to reflect the new structure, ensuring consistency.
