+++ core/chain/tx/hash/index.ts
@@ -2,17 +2,17 @@ import { Chain } from '@core/chain/Chain'
 import { ChainKind, getChainKind } from '@core/chain/ChainKind'
 
 import { DecodedTx } from '../decode'
-import { getCardanoTxHash } from './cardano'
-import { getCosmosTxHash } from './cosmos'
-import { getEvmTxHash } from './evm'
-import { getPolkadotTxHash } from './polkadot'
-import { getRippleTxHash } from './ripple'
-import { getSolanaTxHash } from './solana'
-import { getSuiTxHash } from './sui'
-import { getTonTxHash } from './ton'
-import { getTronTxHash } from './tron'
-import { TxHashResolver } from './TxHashResolver'
-import { getUtxoTxHash } from './utxo'
+import { TxHashResolver } from './resolver'
+import { getCardanoTxHash } from './resolvers/cardano'
+import { getCosmosTxHash } from './resolvers/cosmos'
+import { getEvmTxHash } from './resolvers/evm'
+import { getPolkadotTxHash } from './resolvers/polkadot'
+import { getRippleTxHash } from './resolvers/ripple'
+import { getSolanaTxHash } from './resolvers/solana'
+import { getSuiTxHash } from './resolvers/sui'
+import { getTonTxHash } from './resolvers/ton'
+import { getTronTxHash } from './resolvers/tron'
+import { getUtxoTxHash } from './resolvers/utxo'
 
 const hashHandlers: Record<ChainKind, TxHashResolver<any>> = {
   cardano: getCardanoTxHash,
