+++ core/chain/tx/hash/resolvers/cardano.ts
@@ -1,7 +1,7 @@
 import { Serialization } from '@cardano-sdk/core'
 import { OtherChain } from '@core/chain/Chain'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from '../resolver'
 
 export const getCardanoTxHash: TxHashResolver<OtherChain.Cardano> = ({
   encoded,
