+++ core/chain/tx/hash/resolvers/cosmos.ts
@@ -2,7 +2,7 @@ import { CosmosChain } from '@core/chain/Chain'
 import { stripHexPrefix } from '@lib/utils/hex/stripHexPrefix'
 import { sha256 } from 'viem'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from './resolver'
 
 export const getCosmosTxHash: TxHashResolver<CosmosChain> = ({
   serialized,
