+++ core/chain/tx/hash/resolvers/evm.ts
@@ -1,7 +1,7 @@
 import { EvmChain } from '@core/chain/Chain'
 import { keccak256 } from 'viem'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from './resolver'
 
 export const getEvmTxHash: TxHashResolver<EvmChain> = ({ encoded }) =>
   keccak256(encoded)
