+++ core/chain/tx/hash/resolvers/ripple.ts
@@ -1,7 +1,7 @@
 import { OtherChain } from '@core/chain/Chain'
 import { hashes } from 'xrpl'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from './resolver'
 
 export const getRippleTxHash: TxHashResolver<OtherChain.Ripple> = ({
   encoded,
