+++ core/chain/tx/hash/resolvers/solana.ts
@@ -1,7 +1,7 @@
 import { OtherChain } from '@core/chain/Chain'
 import { shouldBePresent } from '@lib/utils/assert/shouldBePresent'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from './resolver'
 
 export const getSolanaTxHash: TxHashResolver<OtherChain.Solana> = ({
   signatures,
