+++ core/chain/tx/hash/resolvers/ton.ts
@@ -1,6 +1,6 @@
 import { OtherChain } from '@core/chain/Chain'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from './resolver'
 
 export const getTonTxHash: TxHashResolver<OtherChain.Ton> = ({ hash }) =>
   Buffer.from(hash).toString('hex')
