+++ core/chain/tx/hash/resolvers/tron.ts
@@ -2,7 +2,7 @@ import { OtherChain } from '@core/chain/Chain'
 import { stripHexPrefix } from '@lib/utils/hex/stripHexPrefix'
 import { bytesToHex } from 'viem'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from './resolver'
 
 export const getTronTxHash: TxHashResolver<OtherChain.Tron> = ({ id }) =>
   stripHexPrefix(bytesToHex(id))
