+++ core/chain/tx/hash/resolvers/utxo.ts
@@ -1,6 +1,6 @@
 import { UtxoChain } from '@core/chain/Chain'
 
-import { TxHashResolver } from './TxHashResolver'
+import { TxHashResolver } from './resolver'
 
 export const getUtxoTxHash: TxHashResolver<UtxoChain> = ({ transactionId }) =>
   transactionId
